Handle invalid protocol messages, drop connection
Prior to this commit, a client sending invalid messages caused
a crash due to incorrect specification of the Exception class
for handling invalid messages, but also the connection was left
intact.
Since counterparties should never send messages which do not
conform to the syntax of OnionCustomMessage, we now, after this
commit, swallow any such exception with a debug message and
immediately drop the connection to that peer.
This will address for example the common case of someone trying
to visit the onion service at port 80 with a browser.