Change client factory to not be reconnecting
Prior to this commit we were using a ClientService
to set up the connection as client, this meant that
reconnection was continuously attempted when the connection
was lost. This was not the desired behaviour for p2p connections
to makers e.g. Only in the case of a directory node do we want
to have connections be persistent. This commit changes
OnionClientFactory to be derived from protocol.ServerFactory,
and connects to the endpoint directly, meaning that when the connection
is lost no reconnection attempt is made.